services/api-gateway/requirements.txt
Flask==2.2.5
Werkzeug==2.2.2
requests==2.31.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp

